Move common definitions to their own file
This change moves the common definitions to their own file so
that we can later rewrite the driver to accept the structure
as input.

A caveat with this approach is that it does prevent shell scripts
from updating the joystick, however, it eases the burden of
verification in the kernel, and it's not too hard to write
wrapper programs to write kernel structures to the files.