util: Remove generated file from distribution tarball
util_char_map.c is generated at compile time by the Python script
x52_char_map_gen.py from the configuration x52_char_map.cfg. However,
because it was listed in libx52util_la_SOURCES, it was getting added to
the distribution.

In addition, removing it from the distribution caused `make distcheck`
to fail, since the generation script was called relative to the current
directory, instead of relative to `$(srcdir)`. This commit also fixes
that issue.