Delete obsolete man pages
Prior to this change, a separate copy of the man pages for the libx52
functions was being maintained since I couldn't get Doxygen to work
properly at the time (I don't remember, it's been over 5 years).

The man pages are out of date and don't match up with the current
version of libx52, and haven't been for some time. Now that we have
functional documentation generated via Doxygen, it is time to get rid of
the obsolete stuff.

This change deletes the entire man folder, and doesn't change anything
else, since it was self-contained.

[skip ci]