Fix install error on macOS when doxygen is found
The `install` command on macOS does not support the `-D` flag. It is not
required anyway, since the directory components are created by the
previous line.

This commit also updates the options to rsync to not copy in 'archive'
mode. Because it is used in `make install`, it is likely to be run as
root, and therefore, the root user should own the files in the
destination directory.