Disable -Wpacked compiler flag
See issue #31. In systems with a newer version of libusb (1.0.24
onwards), GCC raises a warning that the packed attribute is unnecessary
for `libusb_control_setup`. This is not really a problem for libx52,
since it doesn't call libusb_control_setup directly, but since the
libx52 build framework treats all warnings as errors by default, it will
prevent the build from working.

This change removes the -Wpacked flag from AX_COMPILER_FLAGS, and this
should let the build pass without any issues.