Update hid-saitek-x52 module to use default kernel mapping
Prior to this change, the kernel module was using a customized mapping.
This causes issues between users who rely on the mapping not changing
between the no-module and with-module cases.

This change updates the kernel module to report in the exact same
fashion as the default kernel, without any remapping. While this does
mean that the buttons on the X52 and X52 Pro would report different
values, a userspace daemon can always use libx52io to get the button
mapping regardless of which device is plugged in.